ansible/roles/web_servers/files/student-django/base/serializers.py
from rest_framework.serializers import ModelSerializer
from base.models import Student, Lecturer

class StudentSerializer(ModelSerializer):
    class Meta:
        model = Student
        fields = '__all__'

class LecturerSerializer(ModelSerializer):
    class Meta:
        model = Lecturer
        fields = '__all__'