ansible/roles/web_servers/files/student-django/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
psycopg2==2.9.9
sqlparse==0.4.4
uWSGI==2.0.24
